Phase 4a: extract types and view-config helpers out of WorksCalendar.tsx
- WorksCalendar.types.ts: all exported types (WorksCalendarProps, CalendarApi,
  CalendarView/Role, ScheduleInstantiationLimits, DispatchEvaluator, helper aliases)
- core/calendarViewConfig.ts: ALL_VIEWS, DEFAULT_SCHEDULE_INSTANTIATION_LIMITS,
  viewRange, exportVisibleEvents, opAnnouncement
- WorksCalendar.tsx: re-exports public API from types file; imports from both
  new modules; trimmed date-fns/lucide/import lines
- Net reduction: 1916 → 1591 lines (-325)
